refactor:Make fields of ImageDef public. - fixup references. feat:ImageDef:Add texture coordinates.

+++ include/gfx/BinImageDef.h
@@ -13,7 +13,7 @@ namespace dagui
     class DAGUI_API BinImageDef : public ImageDef
     {
     public:
-        BinImageDef(std::size_t width, std::size_t height, std::size_t numComponents);
+        BinImageDef(std::uint32_t width, std::uint32_t height, std::size_t numComponents);
 
         Image* createImage() const override;
     private:
